requirements.txt
Django==2.0
django-simple-history==2.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
